Move api_exception_handler out of throttling to drop the circular import (#2140)
