+++ config/settings.py
@@ -481,7 +481,7 @@ REST_FRAMEWORK = {
     "DEFAULT_VERSION": "v1",
     "ALLOWED_VERSIONS": ["v1", "v2"],
     "DEFAULT_THROTTLE_CLASSES": ["apps.api.throttling.APIRateThrottle"],
-    "EXCEPTION_HANDLER": "apps.api.throttling.api_exception_handler",
+    "EXCEPTION_HANDLER": "apps.api.exception_handlers.api_exception_handler",
 }
 
 SPECTACULAR_SETTINGS = {
